+++ NEWS.md
@@ -1,5 +1,7 @@
 # diceR 0.1.0.9000
 
+* New dimension reduction method: t-SNE (@dustin21)
+
 * Better annotated progress bar using `progress` package
 
 * Speed up the operation that transforms a matrix to become "NMF-ready"